+++ R/conversion.R
@@ -301,9 +301,6 @@ date_ts2date <- function(date_ts, frequency_ts) {
 #'
 #' ts2df(AirPassengers)
 #'
-#' @keywords internal
-#' @noRd
-#'
 ts2df <- function(x) {
     # Check de l'objet x
     assert_ts(x, .var.name = "x", allow_mts = TRUE)
@@ -319,8 +316,18 @@ ts2df <- function(x) {
 
     if (frequency_ts == 12L) {
         first_date_ts <- as_yyyymm(first_timeunits)
+        date_vec <- seq.Date(
+            from = date_ts2date(first_date_ts, frequency_ts = 12L),
+            length.out = length_series,
+            by = "month"
+        )
     } else if (frequency_ts == 4L) {
         first_date_ts <- as_yyyymm(first_timeunits)
+        date_vec <- seq.Date(
+            from = date_ts2date(first_date_ts, frequency_ts = 4L),
+            length.out = length_series,
+            by = "quarter"
+        )
     }
 
     rownames_libelles <- libelles(
@@ -328,7 +335,8 @@ ts2df <- function(x) {
         frequency_ts = frequency_ts,
         n = length_series
     )
-    x <- data.frame(date = rownames_libelles, x)
+    x <- data.frame(date = date_vec, x)
+    rownames(x) <- rownames_libelles
 
     return(x)
 }